chore: move `push_falcon_div` to stdlib event
