+++ assembly/src/parser/grammar.lalrpop
@@ -55,7 +55,6 @@ extern {
         "push_mtnode" => Token::PushMtnode,
         "push_smtpeek" => Token::PushSmtpeek,
         "push_u64div" => Token::PushU64Div,
-        "push_falcon_div" => Token::PushFalconDiv,
         "and" => Token::And,
         "assert" => Token::Assert,
         "assertz" => Token::Assertz,
@@ -676,7 +675,6 @@ SystemEvent: Instruction = {
     "adv" "." "push_mtnode" => Instruction::SysEvent(SystemEventNode::PushMtNode),
     "adv" "." "push_smtpeek" => Instruction::SysEvent(SystemEventNode::PushSmtPeek),
     "adv" "." "push_u64div" => Instruction::SysEvent(SystemEventNode::PushU64Div),
-    "adv" "." "push_falcon_div" => Instruction::SysEvent(SystemEventNode::PushFalconDiv),
 }
 
 #[inline]
